Three variable function
Added the three variable system of equation function. I added a useTwo
and useThree which are set to “false”. Basically when you select the
number of equations in the HTML, it would change it to true (I don’t
know how that would work). I made so that if it is true, it would run
the function corresponding to the type of equation. I think you should
also add it so that it only runs the function when the button is
pressed, so an && needs to be added to the if at the bottom.