Fractions and removing main.js
I added the fractions to the javascript for both the 2 system of
equations and the 3 system of equations. I also changed the HTML
because it still has “<script src="js/main.js"></script>” but main.js
doesn’t exist. You change it back if it is suppose to be like that.